% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{count}
\alias{count}
\title{Multivariate imputation}
\usage{
count(x, what)
}
\arguments{
\item{x}{Dataframe with missing values}

\item{order}{Imputation order (seq/hire)}

\item{method}{Imputation method (dvinemed/dvinemean/lm/rq/rlm/lqr)
Function used for is.mono}
}
\description{
Multivariate imputation
}
