% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esg_sim}
\alias{esg_sim}
\title{ESG Simulation Dataset}
\format{
A data frame with 1000 rows and 7 columns:
\describe{
\item{X2015, ..., X2021}{Company ESG scores in 2015 to 2021}
}
}
\source{
Simulated data for demonstration purposes
}
\usage{
esg_sim
}
\description{
A simulated dataset containing Environment, Social, and Governance (ESG) scores
for hypothetical companies.
}
\keyword{datasets}
