% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_impute.R
\name{fit_params}
\alias{fit_params}
\title{Univariate imputation}
\usage{
fit_params(fit)
}
\arguments{
\item{fit}{A gamlss object}

\item{x}{Dataframe with missing values}

\item{uscale}{Dataframe is u-scale or x-scale}

\item{print_pair}{If need to print pairs plot
Internal function of getting fitted parameter}
}
\description{
Univariate imputation
}
