% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{multi.impute.dvine}
\alias{multi.impute.dvine}
\title{Multi-variable D-vine imputation}
\usage{
multi.impute.dvine(x, order = "seq", method = "dvinemed")
}
\description{
Multi-variable D-vine imputation
}
