% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{multi.impute.q.dvine}
\alias{multi.impute.q.dvine}
\title{Multi-variable D-vine imputation interval}
\usage{
multi.impute.q.dvine(x, order = "seq", method = "dvinemed", q = 0.5)
}
\description{
Multi-variable D-vine imputation interval
}
