% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{multi.impute.q}
\alias{multi.impute.q}
\title{Multi-variable imputation interval}
\usage{
multi.impute.q(x, order = "seq", method = "dvinemed", q = 0.5)
}
\description{
Multi-variable imputation interval
}
