% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{multi.impute}
\alias{multi.impute}
\title{Multi-variable imputation}
\usage{
multi.impute(x, order = "seq", method = "dvinemed")
}
\arguments{
\item{order}{Imputation order (seq/hire)}

\item{method}{Imputation method (dvinemed/dvinemean/lm/rq/rlm/lqr)}
}
\description{
Multi-variable imputation
}
