% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/validation_metrics.R
\name{rmse}
\alias{rmse}
\title{Error metrics}
\usage{
rmse(a, b)

rmse(a, b)
}
\arguments{
\item{a}{Observed data}

\item{b}{Imputed data
Calculate RMSE}

\item{y}{Dataframe without missing values}

\item{y_005}{5\% percentile imputation}

\item{y_050}{Median imputation}

\item{y_095}{95\% percentile imputation}
}
\description{
Error metrics

Validation metrics
}
