% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_impute.R
\name{uni.impute.dvinemed}
\alias{uni.impute.dvinemed}
\title{D-vine median imputation}
\usage{
uni.impute.dvinemed(x, uscale = FALSE, print_pair = FALSE, alpha = 0.5)
}
\description{
D-vine median imputation
}
