% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_impute.R
\name{uni.impute.lqr}
\alias{uni.impute.lqr}
\title{Robustr quantile regression imputation}
\usage{
uni.impute.lqr(x, print_pair = FALSE, p = 0.5)
}
\description{
Robustr quantile regression imputation
}
