% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_impute.R
\name{uni.impute.q}
\alias{uni.impute.q}
\title{Main imputation interface for interval}
\usage{
uni.impute.q(
  x,
  method = "dvinemed",
  uscale = FALSE,
  print_pair = FALSE,
  q = 0.5
)
}
\description{
Main imputation interface for interval
}
