% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_impute.R
\name{uni.impute}
\alias{uni.impute}
\title{Main imputation interface}
\usage{
uni.impute(x, method = "dvinemed", uscale = FALSE, print_pair = FALSE)
}
\arguments{
\item{method}{Imputation method (dvinemed/dvinemean/lm/rq/rlm/lqr)}
}
\description{
Main imputation interface
}
