% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_impute.R
\name{uni.impute.dvinemed}
\alias{uni.impute.dvinemed}
\title{Univariate imputation}
\usage{
uni.impute.dvinemed(x, uscale = FALSE, print_pair = FALSE, alpha = 0.5)
}
\arguments{
\item{x}{Dataframe with missing values}

\item{uscale}{Dataframe is u-scale or x-scale}

\item{print_pair}{If need to print pairs plot
D-vine median imputation}
}
\description{
Univariate imputation
}
